Avoid letting image previews stay open
If the user hovers an image link but moves away before the image
has had the chance to load, the image preview would stay open until
the user hover the link again and moves away. To avoid that, mark
the image so that if the image finishes loading after the user has
moved away, the image preview won't be created at all.